agents/requirements.txt
aiohttp>=3.9
nats-py>=2.7
